Packages everything
