+++ .gitignore
@@ -16,3 +16,4 @@ bin/
 /text-ui-test/ACTUAL.txt
 text-ui-test/EXPECTED-UNIX.TXT
 src/main/java/Test.java
+src/main/java/Duke/Test.java
