+++ src/main/java/Duke/Deadline.java
@@ -1,5 +1,6 @@
+package Duke;
+
 import java.text.ParseException;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.text.SimpleDateFormat;
