+++ src/main/java/Duke/Task.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.text.SimpleDateFormat;
